refactor preferences
