+++ src/components/CardItemWithActions.js
@@ -1,5 +1,5 @@
 import React, { useState, useContext, useEffect } from 'react'
-import PreferencesContext from '../contexts/PreferencesContext'
+import PreferencesContext from '../preference/PreferencesContext'
 import { BiBookmarkPlus } from "react-icons/bi";
 import { BiBookmarkMinus } from "react-icons/bi";
 import { trackBookmarkFrom, trackUnbookmarkFrom } from "../utils/Analytics"
