+++ src/preference/AppReducer.js
